requirements.txt
Flask
Gunicorn
tweepy
camelcase
python-dotenv